comments fixed
