+++ src/test/java/uz/myproject/chill/constants/Endpoints.java
@@ -1,0 +1,9 @@
+package uz.myproject.chill.constants;
+
+ public class Endpoints {
+        public static final String BASE_URL = "http://qa-scooter.praktikum-services.ru";
+        public static final String CREATE_COURIER = "/api/v1/courier";
+        public static final String LOGIN_COURIER = "/api/v1/courier/login";
+        public static final String DELETE_COURIER = "/api/v1/courier/";
+        public static final String ORDERS = "/api/v1/orders/";
+    }
